nlp-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
torch>=2.2.0
sentence-transformers==2.7.0
transformers==4.36.2
huggingface-hub==0.20.0
numpy>=1.26.2
scikit-learn>=1.3.2
spacy>=3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
requests>=2.31.0
gunicorn==21.2.0
openai>=1.12.0
jobspy>=0.4.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
